aws cloudprovider - separate aws sdk provider from aws manager
